pyproject.toml
[project]
name = "newshackpack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy<2",
    "pyvespa>=0.53.0",
    "torch<2.3.0",
    "transformers>=4.48.3",
]
